fix: a float->double promotion I added broke the ASan build
scaled_add's backward rule passes an f32 coefficient to scaled_add(), which
takes double. Clang rejects the implicit widening under -Wdouble-promotion,
which CMakeLists applies to src/ and the asan preset turns into an error.

This built cleanly here and broke CI. Two reasons it was invisible:

- GCC does not enable -Wdouble-promotion, so the release build never mentioned
  it.
- An INCREMENTAL build does not re-emit warnings for files it did not
  recompile, so every "grep for warning" after a rebuild was reading nothing.

Found by running the ASan job in the CI container rather than trusting the
local build. Verified: `cmake --preset asan` -- the configuration CI uses --
now builds clean under -Werror.

+++ src/autograd/autograd.cpp
@@ -216,9 +216,16 @@ void apply_backward(const NodePtr& np, const Tensor& grad, GradMap& grads) {
             // one. `scale_grad` uses the second operand as a dummy with a zero
             // coefficient: one dispatch rather than the two a scalar multiply
             // would cost, since a scalar operand is materialised as a tensor.
+            //
+            // The cast is explicit because the coefficients are stored as f32 --
+            // that is what the shader receives -- while `scaled_add` takes
+            // double. Clang rejects the implicit widening under
+            // -Wdouble-promotion, which the asan preset turns into an error and
+            // GCC does not enable; this built cleanly here and broke the ASan
+            // job.
             const ScaledAddParams p = node.params.get<ScaledAddParams>();
             const auto scale_grad = [&](float k) {
-                return k == 1.0F ? grad : scaled_add(grad, grad, k, 0.0);
+                return k == 1.0F ? grad : scaled_add(grad, grad, static_cast<double>(k), 0.0);
             };
             accumulate(grads, a, scale_grad(p.alpha));
             accumulate(grads, b, scale_grad(p.beta));
